Fixed double free when using just an environment variable as the command
